requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Deep learning & NLP
torch>=2.0.0
transformers>=4.30.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Excel support
openpyxl>=3.1.0

# Optional: Hyperparameter optimization (used in early experiments)
hyperopt>=0.2.7

opensmile==2.6.0
openai-whisper==20240930
regex==2026.1.15
ffmpeg==7.1.1